Split up module before packaging
