--- README.md
@@ -1,2 +1,0 @@
-# concierge-cli
-Concierge repository projects management CLI.
