+++ concierge_cli/__init__.py
@@ -1,0 +1,8 @@
+"""
+Concierge repository projects management CLI.
+"""
+__author__ = 'VSHN AG'
+__email__ = 'tech@vshn.ch'
+__url__ = 'https://github.com/vshn/concierge-cli'
+__version__ = '0.1.0.dev1'
+__license__ = 'BSD 3-Clause'
