+++ setup.py
@@ -1,19 +1,49 @@
+#!/usr/bin/env python3
 """
 Packaging setup for Concierge CLI.
 See: https://click.palletsprojects.com/en/7.x/setuptools/
 """
-from setuptools import setup
+from os.path import abspath, dirname, join
+from setuptools import find_packages, setup
+
+import concierge_cli as package
+
+
+def read_file(filename):
+    """Get the contents of a file"""
+    here = abspath(dirname(__file__))
+    with open(join(here, filename)) as file:
+        return file.read()
+
 
 setup(
     name='concierge-cli',
-    version='0.99.0.dev0',
-    py_modules=['cli'],
+    version=package.__version__,
+    license=package.__license__,
+    author=package.__author__,
+    author_email=package.__email__,
+    description=package.__doc__.strip(),
+    long_description=read_file('README.rst'),
+    long_description_content_type='text/x-rst',
+    url=package.__url__,
+    packages=find_packages(exclude=['test*']),
+    include_package_data=True,
+    keywords=['cli', 'gitlab', 'management'],
+    classifiers=[
+        'Intended Audience :: Developers',
+        'License :: OSI Approved :: BSD License',
+        'Programming Language :: Python',
+        'Programming Language :: Python :: 3 :: Only',
+        'Programming Language :: Python :: 3.6',
+        'Programming Language :: Python :: 3.7',
+    ],
+    python_requires='>=3.6',
     install_requires=[
         'Click',
     ],
     entry_points={
         'console_scripts': [
-            'concierge-cli = cli:main',
+            'concierge-cli = concierge_cli.cli:main',
         ],
     },
 )
